% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta.dashboard.shiny-package.R
\docType{package}
\name{pacta.dashboard.shiny-package}
\alias{pacta.dashboard.shiny}
\alias{pacta.dashboard.shiny-package}
\title{pacta.dashboard.shiny: RShiny Dashboard for PACTA Results}
\description{
Defines a RShiny dashboard for PACTA results, and serves it.
}
\author{
\strong{Maintainer}: Alex Axthelm \email{aaxthelm@rmi.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]

Other contributors:
\itemize{
  \item RMI \email{PACTA4investors@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
